+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Show the name of variables at the top of the helpful buffer.
 Improved handling of special forms, and prevent users from
 accidentally unbinding special forms.
 
+Ensure we can find (and jump to) the definition of functions even if the source
+buffer has narrowing in effect.
+
 # v0.3
 
 Fixed a crash on autoloaded functions that aren't loaded yet.